Change instructions to restart collector after Log Insights setup
We've seen several customer reports of the collector not picking up a
new Log Insights configuration after a reload. We should identify and
fix the underlying issue in the collector, but in the meantime, we
should change the instructions to indicate a collector restart rather
than a reload. A restart is fairly low-impact and it's worth avoiding
confusion for now.