Merge setup docs (#28)
Move in-app collector setup documentation into docs repo. This lets us
share the same documentation in-app and in the public docs, ensuring
more consistent documentation.

Along with this, translate all documentation into MDX, since that makes
it easier to share and compose documentation, and to support rendering
tweaks for the two host environments.