Drop per-install track Log Insights setup pages
We do not have these for EXPLAIN or basic query performance setup, and
they don't have any useful information--it's just extra pages where we
need to maintain links. Instead, link the main page directly to the
first step in each install track, and add a list of troubleshooting
pages for tracks that include these.