Work around weird Gatsby and/or MDX scoping issue
With this previous import name, another MDX file was somehow being
used (the amazon_rds one with the same filename and import name but in
a totally different directory and file) instead of the correct one.

This does not seem to be happening for other imports with the same
name.