Rework password generation for monitoring user
Because Gatsby does build-time pre-rendering of its MDX pages, we
cannot generate the password on page render. We cannot generate
passwords on-demand, either, because in several documentation pages we
need the same password twice: once in the CREATE USER command, and a
second time in the connection verification command.

Instead, add a layer of indirection to generate a password at the top
level and pass the generated password down to the page content. Also,
avoid generating a random password during the build-time render
entirely (since there's not much point in generating the same random
password for all users), and use the useLayoutEffect hook as
recommended in Gatsby docs [1].

[1]: https://www.gatsbyjs.com/docs/debugging-html-builds/ --
     technically Gatsby recommends useEffect, but useLayoutEffect
     allows us to avoid flashing the server-generated static password
     before it's replaced with a client-generated password.