Refactor Log Insights and Automated Explain setup docs (#34)
Rearrange the documentation here to make it suitable for including in the
pganalyze app as well as the public documentation site. In the app, we
have a lot more context about your database server so we can provide
more specific setup guidance, and verify setup progress (since the
collector will report on this).

This also clarifies and expands on a lot of the setup instructions to make
installation easier, and updates instructions for current Postgres
versions and modern OS distributions.