Update docs to create pganalyze schema if necessary for log-based EXPLAIN
This only happens right now with Heroku Postgres, where we do not
create a separate monitoring user. We've gotten two support tickets
recently asking about the missing schema.

Given that the log-based EXPLAIN docs are generic and not specific to
any provider, we can't easily provide Heroku-specific instructions
here without restructuring the install process (which is probably not
worth it anyway for such a small change). However, IF NOT EXISTS
prevents problems on other platforms and is available back to 9.3.