Fix images after Gatsby v3 upgrade
It looks like the MDX handling changed in such a way that sometimes
breaks the syntax we were using in most places for handling images:

    import imgPath from '../path/to/image'

    <img src={imgPath} />

This does still work in some places, but it's not clear what details
of the MDX structure affect what works and what doesn't.

Instead, create one-off components wrapping images in these MDX files,
which reliably works around the issue.