Fix rendering of self-managed install button in Chrome
Some of the MDX changes in the Vite upgrade led to the explanatory
text rendering out of the box in Chrome. The CSS here is a bit
tortured, but I couldn't figure out a simple way to straighten this
out. For now, just drop the explicit image dimensions, drop the line
break (which does not make much sense in a flexbox context anyway),
and increase the negative margin to have this match current prod.