Fix className handling in SmartAnchor
SmartAnchor has a mostly-innocuous bug right now where it will add a
bogus 'undefined' class name if the className prop is not specified or
set to 'undefined'. This is a bug in its className handling, which was
bespoke and not relying on the classNames package because it seemed
odd for docs to depend on packages without declaring its own
dependencies.

However, given classNames is present in all current consumers, and is
a fairly small and broadly useful package, rely on it instead to avoid
future bugs and improve readability.