Fix escaping in helper function documentation
In two places, we're using regular expressions with backslashes, but
because these are in TypeScript backtick-quoted strings, the
backslashes are interpreted as escapes for the following
characters. Escaped characters with no special meaning render as
themselves, but this means the backslashes are stripped.

Add additional backslashes to escape the backslashes we want to keep.