Introduce dedicated page for Enterprise Server container settings
In the past this information has been scattered across multiple pages,
making it hard to get an overview. Instead create a dedicated page, like
we have for collector settings.

In passing remove all appendixes from the VM-based setup, make instructions
a bit less AWS specific in places (be explicit about when the AWS settings
are needed), and move troubleshooting steps to the troublshooting page.

This also cleans up the Log Insights setup page, which was mostly moved
to the object storage setup instructions and separate collector install
steps, but still needs to exist to be linked from in-app. Modify it to
avoid duplication with the other new pages added in prior commits.

Last, this removes the LDAP_ADMIN_GROUP_CN setting documentation, since
that setting has not been supported in Enterprise Server for a while.