RDS docs: Replace note on track_activity_query_size with track_io_timing (#111)
* RDS docs: Replace note on track_activity_query_size with track_io_timing

There is no argument for raising track_activity_query_size in the
context of pg_stat_statements these days. This is a historic artifact
from the days where pgss kept a limited query text length and tied
that to the track_activity_query_size setting (usually used for
pg_stat_activity).

Instead, mention track_io_timing as something to adjust, as that will
have a direct impact on the usefulness of pg_stat_statements data.