Log Insights: Use consistent log test instructions for containers
This makes sure the log test instructions are aware of whether the
collector is configured from ENV variables (i.e. running as a container),
or in a virtual machine from a config file.

Additionally, this makes the instructions consistent across RDS/GCP/Azure
by using a component instead of copying the same text multiple times.