Add explicit instructions for Google AlloyDB
We've added support in the collector, but had not yet documented that
our Google Cloud SQL instructions also work for AlloyDB, apart from a
few minor details that this change now clarifies.

In passing this makes sure we use "Amazon RDS and Amazon Aurora"
consistently, and adds the "for PostgreSQL" suffix to "Azure Database"
where it was missing (note we technically also support Azure CosmosDB
for PostgreSQL, but we omit this for now due to formatting challenges.