Install overview: Fix styling for Self-Managed secondary title
Previously this was different between the in-app rendering and the
website docs rendering, because the in-app renderer adds an implicit
p tag around the panel text (such as "Self-Managed"), but the website
does not. Generally this doesn't matter, but in the case of the
self-managed panel it caused a styling difference on the website.