Enterprise EKS setup: Use Entrypoint script, db:structure:load (#262)
Since the v2024.04.0 release, we need to ensure all paths are set up
correctly, not just RVM. Since the Entrypoint has been modified to
allow calling it directly more easily, switch the K8S documentation
to call it to initialize the environment and run rake tasks.

In passing change the use of "db:setup" (which runs the migrations)
to "db:structure:load" (which loads structure.sql) + "db:seed". This
matches the VM-based setup instructions.