CopyCodeBlock: Fix highlighting so it actually works
This didn't work before since neither highlight.js or Prism triggered
for the contents of the React component.

Since we mainly use highlight.js (short of Gatsby's remark plugin for
Prism), highlight CopyCodeBlock contents explicitly using highlight.js.

In passing, also use the pgsql language when "sql" is asked for, so we
support the keywords we commonly use correctly.